16.10 commit

+++ app/Http/Controllers/GroupController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 use App\Models\Group;
 use App\Http\Requests\GroupRequest;
 
+use Gate;
+
 class GroupController extends Controller
 {
     protected $groups;
@@ -32,6 +34,10 @@ class GroupController extends Controller
 
     public function store(GroupRequest $request)
     {
+        if (GroupController::checkAuth()) {
+            return back()->with(['message' => 'У вас нет прав']);
+        }
+
         Group::create($request->all());
 
         return redirect('groups/');
@@ -44,6 +50,10 @@ class GroupController extends Controller
 
     public function update(GroupRequest $request, Group $group)
     {
+        if (GroupController::checkAuth()) {
+            return back()->with(['message' => 'У вас нет прав']);
+        }
+
         $group->update($request->all());
 
         return redirect('groups/');
@@ -51,6 +61,10 @@ class GroupController extends Controller
 
     public function destroy(Group $group)
     {
+        if (GroupController::checkAuth()) {
+            return back()->with(['message' => 'У вас нет прав']);
+        }
+
         $group->delete();
 
         return redirect('groups/');
@@ -67,5 +81,10 @@ class GroupController extends Controller
                 $avg_groups[$group->id][$mark->first()->subject->subject_name] = $mark->avg('mark');
             }
         }
+    }
+
+    public function checkAuth()
+    {
+        return Gate::denies('editEntity', Group::class);
     }
 }
