+++ app/Http/Controllers/MarkController.php
@@ -8,6 +8,8 @@ use App\Models\Student;
 use App\Models\Mark;
 use App\Models\Subject;
 
+use Gate;
+
 class MarkController extends Controller
 {
     public function create(Student $student)
@@ -19,6 +21,10 @@ class MarkController extends Controller
 
     public function store(Request $request, Student $student)
     {
+        if (MarkController::checkAuth()) {
+            return back()->with(['message' => 'У вас нет прав']);
+        }
+
         foreach ($request->Create as $subject_id => $subject_marks) {
             foreach ($subject_marks as $mark) {
                 if ($mark > 1) {
@@ -38,6 +44,10 @@ class MarkController extends Controller
 
     public function update(Request $request, Student $student, Mark $marks)
     {
+        if (MarkController::checkAuth()) {
+            return back()->with(['message' => 'У вас нет прав']);
+        }
+
         foreach ($request->Update as $mark_id => $mark) {
            Mark::where('id', $mark_id)->update(['mark' => $mark]);
         }
@@ -47,8 +57,17 @@ class MarkController extends Controller
 
     public function destroy(Student $student, Mark $mark)
     {
+        if (MarkController::checkAuth()) {
+            return back()->with(['message' => 'У вас нет прав']);
+        }
+
         $mark->delete();
 
         return redirect('students/' . $student->id);
+    }
+
+    public function checkAuth()
+    {
+        return Gate::denies('editEntity', Mark::class);
     }
 }
